Pull hostpolicy resolving/loading out of fx_muxer.cpp
Split reading config for app/lib into separate function from
loading/executing runtime
